Fix runtime. PH: padams
The getter really needs to be called getFoo() ; this is counter to the
style of every other language, but hey, we're a Java shop.

Make the projectfile xbm refer to the right class and remove fileStatusChange
since I don't see where it should come from in the Java code.

Comment out the unimplemented classes from the hibernate.cfg, since
it causes hibernate to bail on startup. Add them back once the Java
implementation is done as well.