[ For those using Maven ]
I disabled the use of Maven snapshot repositories (for Maven plugins),
since:
1. They are not as reliable as expected.
2. Makes Maven download plugins' snapshots quite often.

If you've detected that a plugin retrieved from a snapshot repo is not 
working as expected, there are two ways to solve the problem:
1. Delete the affected plugin from your user's remo (usually ~/.m2). A   
stable version of the plugin will be downloaded next time you call 
Maven. 
2. Delete the complete user's repo. This way next time you call Maven a 
complete repo will be recreated using the stable version of the plugins.