The source updater now: 1. Checkouts the selected project revisions locally (/var/tmp/fds/<project>). 2. By using the local checkout some file meta-data can be retrieved as well (file size only). 3. Most of the required file meta-data fields are strictly file system dependent and I don't see a common way for retrieving it. 4. The whole source update process is now EXTREMELY slow (... each revision is checked into a separate folder instead of being updated, but if I remember correctly that was the plan too).
Note: I'm not sure, if this is the right place (Ade/Giorgos ?) for a revision checkout, but it seems like the right place for a files meta-data creation (and it needs access to those files).

P.S.: At least you can (probably) run some metrics on top of this checkouts too.