A metric's *code* should specify which configuration properties it has (although this might be done through its manifest as well); add a getConfigurationSchema() method that returns the schema as a set of name,type pairs. This is not very sophisticated but we might be able to get away with it for now.
What will happen is that we can apply the configuration found in plugins.xml to the metrics by intersecting the names from getConfigurationScheme() with what is in the config file (possibly we can dispense with types and hand off XML elements instead). This also means we can ask the metric 'what are your configuration values' and render an HTML form to do it.