Improved example data.
Please note that the login names for the webinterface
included in this db dump have changed slightly.
Previously, they were "'Fluffy Bunny'" (so with single
quotes). Those single quotes are now removed.

There are not evaluated metrics for a certain project
right now. I'm investigating if that's in code or in
the sample data.