maybe it's my version of MSYS that sucks, but sh wasn't happy with the semicolons in the classpath, and the build was failing. Quoting the classpath argument solved it for me, I hope it won't break it on other platforms.
ps1: use $(JAVA_CMD) instead of java, since it's there
ps2: I guess the classpath should be built using the build_classpath macro in Makefile.common, but I didn't manage to (read: 'didn't feel like wasting time on') getting it to work. If anybody wants to give it a try, be my guest.