Whitespace +
In the updaters for data sources we *must* make sure that run()
finishes with a call to removeUpdater() so that the core
updater knows run() is done. I don't know if there's a better
way to query job existence or status -- scheduler dudes?

Fixed several code paths where a removeUpdater() was missed,
leading to the updater thinking projects and updaters were
busy forever.