When a directory is deleted:
- we only get the delete notification for the directory
  in the list of changed paths, not the whole subtree.
- the node for the directory isn't listed as a directory

So we need to check each deleted node if it *was* a directory,
and if so we will need to treat it as a directory. After that we
need to add entries for all of the files underneath the directory,
explicitly deliting them as well. We probably want to do that
depth-first.

I couldn't find a method in ProjectFile right now that does the
'list of files in this directory' right now, but there is lots
of yucky cruft. Will try to avoid distraction by needed cleanup
when implementing 'ls'.