Today's fundamental flaw:
When going 'back' a revision from a ProjectVersion, do
remember to stay within the same project, otherwise the
previous revision to r.6 of project sport might be r.9
of project yoyo, which is kind of confusing.

PS1: why isn't this also a member function? Is it a bean thing?
PS2: i won't promise this is the only fundamental flaw for today.