WebUI::Files:: Implementation of a simple file browser (needs a further polishing though).
How to use - a click on a folder will switch to that folder, or click on the "Top" or "Previous" link to move back in the folders tree.

P.S.: Currently there is a small bug, so when you switch to another project you'll still see the files of the old project. I'll fix that as next.